+++ CHANGES.md
@@ -5,6 +5,7 @@
 
 - Added bulk import tests.
 - Added pydantic to dependencies.
+- Added validation on the metadata and srcdocdata input of the uploadtask endpoint.
 
 
 ## 0.8 (2024-04-18)